t1309: use a non-loaded branch name in the `onbranch` test cases
The `onbranch` test cases in question do not actually want to include
anything; Instead, they want to verify that the `onbranch` code path
does not regress in the early-config case or in the non-Git case, where
the `onbranch` include is actually ignored.

Therefore, the actual branch name does not matter at all. We might just
as well avoid racially-charged names here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>