t4015: let the test pass with any default branch name
We do not need to hard-code the actual branch name, as we have access to
the reflog and can be a lot more precise at what we want.

Strangely enough, this test case would have succeeded even with an
overridden default branch name, obviously for the wrong reason. Let's
verify that it passes for the expected reason, by looking for a
tell-tale in Git's output.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>