config: plumb --literal-value into config API
The git_config_set_multivar_in_file_gently() and related methods now
take a 'flags' bitfield, so add a new bit representing the
--literal-value option from 'git config'. This alters the purpose of the
value_regex parameter to be an exact string match. This requires some
initialization changes in git_config_set_multivar_in_file_gently() and a
new strcmp() call in the matches() method.

The new CONFIG_FLAGS_LITERAL_VALUE flag is initialized in
builtin/config.c based on the --literal-value option, and that needs to
be updated in several callers.

This patch only affects some of the modes of 'git config', and the rest
will be completed in the next change.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>