commit: fix stdin conflict between message and hook
If git commit is executed with -F - (meaning read the commit message
from stdin), and pre-commit hook is also reading from stdin, the
message itself was consumed by the hook before reaching the point
where it is read for the commit message.

Fix this by detecting this case, and passing this information to
run_commit_hook.

Signed-off-by: Orgad Shaneh <orgads@gmail.com>