upload-pack: kill pack-objects helper on signal or exit
We spawn an external pack-objects process to actually send objects to
the remote side. If we are killed by a signal during this process,
then pack-objects may continue to run. As soon as it starts producing
output for the pack, it will see a failure writing to upload-pack and
exit itself. But before then, it may do significant work traversing
the object graph, compressing deltas, etc, which will all be
pointless. So let's make sure to kill as soon as we know that the
caller will not read the result.

Signed-off-by: Junio C Hamano <gitster@pobox.com>