t7817: do not depend on any specific default branch name
Elsewhere, we introduced support to override the default branch name in
the tests via `GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME`. Therefore, we
should avoid expecting a particular hard-coded name in test scripts.

So let's rename the initial branch immediately to `main` and work with
that.

This avoids breakages in t7817 when the default branch name is changed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>