diffcore-rename: change a few comments to use 'add' instead of 'create'
As with the last commit, it feels odd to talk about "creation" (which to
me implies a brand new file, i.e. one not associated with an existing
file, and thus one that could not have been a rename) when we are using
the files to look for renames.  Use the term "addition" instead.

Signed-off-by: Elijah Newren <newren@gmail.com>