mergetool: break setup_tool out into separate initialization function
This is preparation for the following commit where we need to source the
mergetool shell script to look for overrides before `run_merge_tool` is
called. Previously `run_merge_tool` both sourced that script and invoked
the mergetool.

In the case of the following commit, we need the result of the
`automerge_enabled` override, if it exists, well before we actually run
`run_merge_tool`.

A new function `initialize_merge_tool` was chosen for consistency with
`run_merge_tool` since `setup_tool` and `setup_user_tool` are not
exposed or called directly.

Signed-off-by: Seth House <seth@eseth.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>