cache-tree: avoid path comparison loop when silent
The verify_cache() method is used both for debugging issues with the
cached tree extension but also to avoid using the extension when there
are unmerged entries. It also checks for cache entries out of order with
respect to file-versus-directory sorting.

In 996277c (Refactor cache_tree_update idiom from commit, 2011-12-06),
the silent option was added to remove the progress indicators from the
initial loop looking for unmerged entries. This was changed to be a flag
in e859c69 (cache-tree: update API to take abitrary flags, 2012-01-16).

In both cases, the silent option is ignored for the second loop that
checks for file-versus-directory sorting. It must be that this loop is
intended only for debugging purposes and is not actually helpful in
practice.

Since verify_cache() is called in cache_tree_update(), which is run
during 'git commit', we could speed up 'git commit' operations by not
iterating through this loop another time. Of course, noticing this loop
requires an incredibly large index.

To get a measurable difference, I needed to run this test on the
Microsoft Office monorepo, which has over three million entries in its
index. I used 'git commit --amend --no-edit' as my command and saw the
performance go from 752ms to 739ms with this change.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>