update-index: drop the_index, the_repository
To reduce the need for the index compatibility macros, we will replace
their uses in update-index mechanically. This is the most interesting
change, which creates global "repo" and "istate" pointers. The macros
can then be mechanically replaced by instances that use the istate
pointer instead of the version that autocompletes to use the_index.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>